+++ linuxkm/include.am
@@ -13,8 +13,10 @@ EXTRA_DIST += m4/ax_linuxkm.m4 \
 	      linuxkm/pie_last.c \
 	      linuxkm/linuxkm_memory.c \
 	      linuxkm/linuxkm_wc_port.h \
+	      linuxkm/x86_vector_register_glue.c \
 	      linuxkm/lkcapi_glue.c \
+	      linuxkm/lkcapi_aes_glue.c \
+	      linuxkm/lkcapi_sha_glue.c \
 	      linuxkm/lkcapi_ecdsa_glue.c \
 	      linuxkm/lkcapi_ecdh_glue.c \
-	      linuxkm/lkcapi_rsa_glue.c \
-	      linuxkm/x86_vector_register_glue.c
+	      linuxkm/lkcapi_rsa_glue.c
